+++ src/super/i-data/i-data.ts
@@ -1,3 +1,5 @@
+/* eslint-disable max-lines */
+
 /*!
  * V4Fire Client Core
  * https://github.com/V4Fire/Client
@@ -12,6 +14,8 @@
  */
 
 import symbolGenerator from 'core/symbol';
+import SyncPromise from 'core/promise/sync';
+
 import { deprecate, deprecated } from 'core/functools/deprecation';
 
 import RequestError from 'core/request/error';
@@ -118,7 +122,7 @@ export default abstract class iData extends iBlock implements iProgress {
 	readonly dataProvider?: string;
 
 	/**
-	 * Initial parameters for a data provider instance
+	 * Initial parameters for the data provider instance
 	 */
 	@prop({type: Object, required: false})
 	readonly dataProviderOptions?: ProviderOptions;
@@ -164,6 +168,22 @@ export default abstract class iData extends iBlock implements iProgress {
 	 */
 	@prop({type: [Boolean, Function]})
 	readonly defaultRequestFilter?: RequestFilter;
+
+	/**
+	 * If true, all requests to the data provider are suspended till you don't manually force it.
+	 * This prop is used when you want to organize the lazy loading of components.
+	 * For instance, you can load only components in the viewport.
+	 */
+	@prop(Boolean)
+	readonly suspendRequestsProp: boolean = false;
+
+	/**
+	 * Enables the suspending of all requests to the data provider till you don't manually force it.
+	 * Also, the parameter can contain a promise resolve function.
+	 * @see [[iData.suspendRequestsProp]]
+	 */
+	@system((o) => o.sync.link())
+	suspendRequests?: boolean | Function;
 
 	/**
 	 * @deprecated
@@ -257,7 +277,7 @@ export default abstract class iData extends iBlock implements iProgress {
 	}
 
 	/**
-	 * Request parameters for a data provider.
+	 * Request parameters for the data provider.
 	 * Keys of the object represent names of data provider methods.
 	 * Parameters that associated with provider methods will be automatically appended to
 	 * invocation as parameters by default.
@@ -306,6 +326,15 @@ export default abstract class iData extends iBlock implements iProgress {
 	 */
 	@system()
 	protected dp?: Provider;
+
+	/**
+	 * Unsuspend requests to the data provider
+	 */
+	unsuspendRequests(): void {
+		if (Object.isFunction(this.suspendRequests)) {
+			this.suspendRequests();
+		}
+	}
 
 	/** @override */
 	initLoad(data?: unknown, opts: InitLoadOptions = {}): CanPromise<void> {
@@ -727,7 +756,7 @@ export default abstract class iData extends iBlock implements iProgress {
 	}
 
 	/**
-	 * Initializes component data from a data provider.
+	 * Initializes component data from the data provider.
 	 * This method is used to map `db` to component properties.
 	 * If the method is used, it must return some value that not equals to undefined.
 	 */
@@ -877,7 +906,7 @@ export default abstract class iData extends iBlock implements iProgress {
 	}
 
 	/**
-	 * Returns default request parameters for the specified method
+	 * Returns default request parameters for the specified data provider method
 	 * @param method
 	 */
 	protected getDefaultRequestParams<T = unknown>(method: string): CanUndef<DefaultRequest<T>> {
@@ -964,6 +993,26 @@ export default abstract class iData extends iBlock implements iProgress {
 	}
 
 	/**
+	 * Returns a promise that will be resolved when the component can produce requests to the data provider
+	 */
+	protected waitPermissionToRequest(): Promise<boolean> {
+		if (this.suspendRequests === false) {
+			return SyncPromise.resolve(true);
+		}
+
+		return this.async.promise(() => new Promise((resolve) => {
+			this.suspendRequests = () => {
+				resolve(true);
+				this.suspendRequests = false;
+			};
+
+		}), {
+			label: $$.waitPermissionToRequest,
+			join: true
+		});
+	}
+
+	/**
 	 * Creates a new request to the data provider
 	 *
 	 * @param method - request method
@@ -984,11 +1033,20 @@ export default abstract class iData extends iBlock implements iProgress {
 			reqParams = Object.reject(opts, asyncFields),
 			asyncParams = Object.select(opts, asyncFields);
 
-		const
-			req = this.async.request<RequestResponseObject<D>>((<Function>this.dp[method])(body, reqParams), asyncParams),
-			is = (v) => v !== false;
+		const req = this.waitPermissionToRequest()
+			.then(() => {
+				if (this.dp == null) {
+					throw new ReferenceError('The data provider to request is not defined');
+				}
+
+				const rawRequest = (<Function>this.dp[method])(body, reqParams);
+				return this.async.request<RequestResponseObject<D>>(rawRequest, asyncParams);
+			});
 
 		if (this.mods.progress !== 'true') {
+			const
+				is = (v) => v !== false;
+
 			if (is(opts.showProgress)) {
 				void this.setMod('progress', true);
 			}